parte 2 da revolução: codando direito pra obedecer o linter
